+++ CHANGELOG.md
@@ -10,5 +10,8 @@
 - command to include a stylesheet `[!stylesheet]: file.css`
 - option to export to pdf `--pdf`
 - option to watch (and export to html) `--watch`
+- option to bundle everything in one file `--bundle`
 - stylesheets supporting math and code highlighting (with highlightjs)
-- auto base64 converting for images for a standalone html
+- auto base64 converting for images for a standalone html when bundling flag is set
+- support for `mdconf.json`config file
+- MathJax script to html